requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
pytz==2024.1
pyserial==3.5